+++ API/11st/product_search/services/request.service.ts
@@ -1,0 +1,7 @@
+import * as rp from 'request-promise';
+import { APISite } from '../src/option';
+import { SiteRequest } from '../interface/SiteRequest.interface';
+
+export const requestService = async (qs: SiteRequest) => {
+    return rp.get({ url: APISite, qs, json: true, encoding: null });
+};
