+++ API/common/interface/index.ts
@@ -1,0 +1,1 @@
+export * from './ProductScrapResult.interface';
