--- API/tsconfig.json
@@ -1,12 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "es5",
-        "module": "commonjs",
-        "moduleResolution": "node",
-        "declaration": false,
-        "noImplicitAny": true,
-        "allowJs": true,
-        "removeComments": true
-    },
-    "exclude": ["node_modules"]
-}
