+++ express-for-typescript/src/product/callApp.service.ts
@@ -1,5 +1,5 @@
 import { ProductRequest } from './Interface/productRequest.interface';
-var api_11st = require('../../../API/11st/dist/src/index.js');
+var api_11st = require('../../../API/11st/dist/product_search/src/index.js');
 
 export const callApp = async (data: ProductRequest) => {
     try {
